Support #19311 - hostnames
This adds an additional Docker network with hostnames for each component - dina.local for the UI and api.dina.local for Object Store API. This allows the local deployment to more closely resemble the eventual Openshift version. To use these hostnames outside of the Docker network (i.e. on the host machine), corresponding entries must be added to the hosts file as outlined in the readme.